feat: add basic game functionality
At this point, the cards properly update state when clicked for the
first time, current and best score are properly tracked and updated, and
the GameDisplay is rebuilt from scratch on a loss, resetting the state
of all Card components. Next steps are retreiving the images for display
and creating a function to randomize placements on each Card click.